requirements.txt
aiogram
flask
python-dotenv
solana
pydantic
sqlalchemy
jinja2
requests
